feat: move e2e tests
