+++ test/e2e/cases/adopted-probes.test.ts
