+++ test/e2e/cases/adoption-code.test.ts
