+++ test/e2e/cases/location.test.ts
