+++ wallaby.e2e.js
@@ -6,10 +6,10 @@ export default function wallaby () {
 			'public/v1/*',
 			'public/**/*.yaml',
 			'test/plugins/**/*',
-			'test/tests/e2e/client.ts',
-			'test/tests/e2e/setup.ts',
-			'test/tests/e2e/utils.ts',
-			'test/tests/e2e/docker.ts',
+			'test/e2e/client.ts',
+			'test/e2e/setup.ts',
+			'test/e2e/utils.ts',
+			'test/e2e/docker.ts',
 			'src/**/*.ts',
 			'migrations/*',
 			'seeds/**/*',
@@ -17,12 +17,12 @@ export default function wallaby () {
 			'package.json',
 		],
 		tests: [
-			'test/tests/e2e/**/*.test.ts',
+			'test/e2e/**/*.test.ts',
 		],
 
 		setup (w) {
 			const path = require('path');
-			w.testFramework.addFile(path.resolve(process.cwd(), 'test/tests/e2e/setup.js'));
+			w.testFramework.addFile(path.resolve(process.cwd(), 'test/e2e/setup.js'));
 			w.testFramework.timeout(20000);
 		},
